requirements.txt
Python-CartolaFC
pyodbc